tf-serving: ensure modelkit.testing is importable without tf serving
when tensorflow-serving-api is not present importing anything from
modelkit.core.models.tensorflow_model is not possible
catch NameErrors when importing from modelkit.testing
